Use primary key from DB as alarm id.
git-svn-id: https://kraigsandroid.googlecode.com/svn/trunk@49 81ac7b4a-400a-11de-9cfd-b5a970547080

+++ android/alarmclock/src/com/angrydoughnuts/android/alarmclock/AlarmBroadcastReceiver.java
@@ -20,7 +20,7 @@ public class AlarmBroadcastReceiver extends BroadcastReceiver {
   @Override
   public void onReceive(Context context, Intent recvIntent) {
     Bundle input = recvIntent.getExtras();
-    int alarmId = input.getInt("task_id");
+    long alarmId = input.getLong("task_id");
 
     if (wakeLock == null) {
       PowerManager powerManager =
